Add pgsql step for using custom gucs
Files use the naming scheme of TEST.BRANCH.conf for test (e.g. dbt2) and branch
(e.g. master, REL_16_STABLE) specific postgres configuration parameters.

The ForceScheduler results in a build request where the branch is not set, thus
Git uses HEAD.  Set the default branch to be "master" when the branch is not
specified.  Then we don't need a separate config for HEAD...